:recycle: lobby code
